Оптимизации по проекту
