+++ .gitignore
@@ -1,0 +1,5 @@
+# Created by .ignore support plugin (hsz.mobi)
+!.gitignore
+.*
+.class
+*.jar
